airflow/deltalake_requirements.txt
minio==7.1.16
PyYAML==6.0.1
deltalake==0.10.2
numpy<=1.26.0
pandas<=2.1.0
delta-spark==3.1.0